refactor: Extract each RefreshStrategy to its own file
