+++ app/src/main/java/com/infomaniak/mail/data/cache/mailboxContent/refreshStrategies/InboxRefreshStrategy.kt
@@ -1,0 +1,37 @@
+/*
+ * Infomaniak Mail - Android
+ * Copyright (C) 2025 Infomaniak Network SA
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+package com.infomaniak.mail.data.cache.mailboxContent.refreshStrategies
+
+import com.infomaniak.mail.data.cache.mailboxContent.ImpactedFolders
+import com.infomaniak.mail.data.cache.mailboxContent.ThreadController
+import com.infomaniak.mail.data.models.Folder
+import com.infomaniak.mail.data.models.thread.Thread
+import io.realm.kotlin.TypedRealm
+
+val inboxRefreshStrategy = object : DefaultRefreshStrategy {
+    override fun queryFolderThreads(folderId: String, realm: TypedRealm): List<Thread> {
+        return ThreadController.getInboxThreadsWithSnoozeFilter(withSnooze = false, realm = realm)
+    }
+
+    override fun otherFolderRolesToQueryThreads(): List<Folder.FolderRole> = listOf(Folder.FolderRole.SNOOZED)
+
+    override fun addFolderToImpactedFolders(folderId: String, impactedFolders: ImpactedFolders) {
+        impactedFolders += folderId
+        impactedFolders += Folder.FolderRole.SNOOZED
+    }
+}
